[Feat] [GGFE-162] 상점, 재화정책 레이아웃 구성
