+++ components/admin/purchaseHistory.tsx/MegaphoneList.tsx
@@ -40,6 +40,11 @@ function MegaphoneList () {
                             ))}
                         </TableRow>
                     </TableHead>
+                    <TableBody>
+                        <TableRow>
+                            <TableCell>비어있습니다</TableCell>
+                        </TableRow>
+                    </TableBody>
                 </Table>
             </TableContainer>
