+++ pages/admin/coin.tsx
@@ -1,0 +1,7 @@
+export default function Coin() {
+    return (
+        <>
+
+        </>
+    )
+}
