+++ styles/admin/purchaseHistory/MenuTab.module.scss
@@ -1,0 +1,11 @@
+.top {
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.tab {
+    width: 50%;
+    height: 100%;
+}
